[MatEditor] Category Grouping now works
Rename HlmsJsonListener::saveDatablock to canSaveDatablock to avoid
override name clash

+++ OgreMain/src/OgreScriptTranslator.cpp
@@ -9504,8 +9504,8 @@ namespace Ogre{
                         uint32 numMiplevels = 1u;
                         if( getUInt( *it0, &mipLevelStart ) && getUInt( *it1, &numMiplevels ) )
                         {
-                            passDepthCopy->mMipLevelStart = mipLevelStart;
-                            passDepthCopy->mNumMiplevels = numMiplevels;
+                            passDepthCopy->mMipLevelStart = uint8_t( mipLevelStart );
+                            passDepthCopy->mNumMiplevels = uint8_t( numMiplevels );
                         }
                         else
                         {
